Allow managing multiple emoji collections
